requirements.txt
discord.py==1.3.3
phue==1.1
pywizlight==0.3.2
